pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "protobuf==5.29.3",
    "websocket-client==1.8.0",
    "paho-mqtt==2.1.0",
    "python-dotenv==1.0.1",
    "xmltodict>=0.14.2",
    "aiohttp>=3.11.14",
]
